requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.5
django-oauth-toolkit==2.0.0
djangorestframework==3.13.1
drf-social-oauth2==1.2.1
ecdsa==0.17.0
idna==3.3
jwcrypto==1.3.1
oauthlib==3.2.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.4.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
urllib3==1.26.9
wrapt==1.14.1
